+++ stealth/util/data.cc
@@ -1,8 +1,8 @@
 #include "util.ih"
 
 bool     Util::s_keepAlive = false;
-unsigned Util::s_repeatInterval = 0;
-unsigned Util::s_delayInterval = 0;
+size_t Util::s_repeatInterval = 0;
+size_t Util::s_delayInterval = 0;
 Selector Util::s_selector;
 string   Util::s_runFilename;
 FILE    *Util::s_runFILE = 0;
