+++ stealth/util/lockrunfile.cc
@@ -32,7 +32,7 @@ try
     {
         debug() << "attempting non-blocking mode lock on FD " << 
                                                     fileno(s_runFILE) << endl;
-        for (unsigned idx = 0; idx < s_maxBlockAttempts; ++idx)
+        for (size_t idx = 0; idx < s_maxBlockAttempts; ++idx)
         {
             if (flock(fileno(s_runFILE), LOCK_EX  | LOCK_NB) == 0)
                 throw true;
